fix(transformers): honor dtype and retain torch_dtype compatibility (#2379)
* feat(recipes): default fp32 master weights for full-param torch.optim training

Wire resolve_storage_dtype into the llm, vlm, seq-classification, and
bi-encoder retrieval recipes so full-parameter training with a torch.optim
optimizer defaults model.torch_dtype to float32 (parameters act as the fp32
master copy). No-op for PEFT, TE FusedAdam, or an explicit model.torch_dtype.


* feat(diffusion): use fp32 master weights for full-param finetune configs

Set torch_dtype: float32 (fp32 master) + compute_dtype: bfloat16 (FSDP compute)
on the full-parameter diffusion finetune configs, matching wan2_1_t2v_flow.


* feat(recipes): use fp32 master weights for EAGLE draft training

Store the trainable EAGLE-1/EAGLE-3 draft in fp32 (storage_dtype) so
torch.optim.AdamW keeps an fp32 master copy, and run forward/backward in
bf16 on GPU via a _compute_autocast() context (no FSDP MixedPrecisionPolicy
since the draft is not sharded). Autocast is applied in the train/val loops,
the shared _forward_batch helper, and the P-EAGLE partitioned segment step.


* docs(mixed-precision): document fp32-master default behavior

Add a Default Behavior section explaining that full-parameter torch.optim
training now defaults model.torch_dtype to float32 (with PEFT / TE FusedAdam /
explicit-dtype no-ops), restore the bf16 Adam-state subsection (fp32 master +
bf16 EMA, as used by DeepSeek-V3), update the Risky Pattern section to reflect
that bf16 storage is now opt-in, and update the finetune guide note.


* refactor(config): resolve storage dtype without mutation


* docs(skills): guide explicit fp32 storage for full training


* docs(skills): trim repeated precision guidance


* docs: apply technical writer feedback to training guides


* Attach NVSkills validation signatures


---------

+++ nemo_automodel/_transformers/model_init.py
@@ -995,7 +995,7 @@ def _init_model_bnb_streaming(
     with no_init_weights(), init_empty_weights():
         model = cls._from_config_parent_class(
             hf_config,
-            torch_dtype=torch_dtype,
+            dtype=torch_dtype,
             attn_implementation=attn_implementation,
         )
 
@@ -1239,7 +1239,7 @@ def __init_model(
                 model = cls._from_pretrained_parent_class(
                     pretrained_model_name_or_path,
                     *model_args,
-                    torch_dtype=torch_dtype,
+                    dtype=torch_dtype,
                     attn_implementation=attn_implementation,
                     **kwargs,
                 )
@@ -1256,7 +1256,7 @@ def __init_model(
             model = cls._from_config_parent_class(
                 hf_config,
                 *model_args,
-                torch_dtype=torch_dtype,
+                dtype=torch_dtype,
                 attn_implementation=attn_implementation,
                 **kwargs,
             )
@@ -1347,7 +1347,7 @@ def __init_model(
             model = cls._from_pretrained_parent_class(
                 pretrained_model_name_or_path,
                 *model_args,
-                torch_dtype=torch_dtype,
+                dtype=torch_dtype,
                 attn_implementation=attn_implementation,
                 **kwargs,
             )
@@ -1364,7 +1364,7 @@ def __init_model(
         model = cls._from_config_parent_class(
             hf_config,
             *model_args,
-            torch_dtype=torch_dtype,
+            dtype=torch_dtype,
             attn_implementation=attn_implementation,
             **kwargs,
         )
